added authorization for private apis

+++ controllers/contactController.js
@@ -3,16 +3,16 @@ const Contact = require("../models/contactModel");
 
 //@desc Get all contacts
 //@route GET /api/contacts
-//@access public
+//@access private
 
 const getContacts = asyncHandler(async (req, res) => {
-    const contacts = await Contact.find();
+    const contacts = await Contact.find({ user_id: req.user.id });
     res.status(200).json(contacts);
 });
 
 //@desc Create new contact
 //@route POST /api/contacts
-//@access public
+//@access private
 
 const createContact = asyncHandler(async (req, res) => {
     const { name, email, phone } = req.body;
@@ -22,6 +22,7 @@ const createContact = asyncHandler(async (req, res) => {
     }
 
     const contact = await Contact.create({
+        user_id: req.user.id,
         name,
         email,
         phone,
@@ -31,7 +32,7 @@ const createContact = asyncHandler(async (req, res) => {
 
 //@desc Get contact
 //@route Get /api/contacts/:id
-//@access public
+//@access private
 
 const getContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id);
@@ -44,7 +45,7 @@ const getContact = asyncHandler(async (req, res) => {
 
 //@desc Update contact
 //@route PUT /api/contacts/:id
-//@access public
+//@access private
 
 const updateContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id);
@@ -53,19 +54,28 @@ const updateContact = asyncHandler(async (req, res) => {
         throw new Error("Contact not found");
     }
 
+    if(contact.user_id.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error("User won't have permission to update other user contacts");
+    }
+
     const updatedContact = await Contact.findByIdAndUpdate(req.params.id, req.body, { new: true });
     res.status(200).json(updatedContact);
 });
 
 //@desc Delete contact
 //@route DELETE /api/contacts/:id
-//@access public
+//@access private
 
 const deleteContact = asyncHandler(async (req, res) => { 
     const contact = await Contact.findById(req.params.id);
     if (!contact) {
         res.status(404);
         throw new Error("Contact not found");
+    }
+    if(contact.user_id.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error("User won't have permission to delete other user contacts");
     }
     await Contact.findByIdAndDelete(req.params.id);
     res.status(200).json(contact);
